# refactor(csv): switch csv to use the arrow table buffer directly