The csv iterator has been switched to it builds arrow buffers directly
instead of using the `ColListTableBuilder`. This makes it more efficient
since it will just allocate the types as it is built instead of building
it once and then copying it.

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written