@jsternberg 2019-09-23
Here's the benchmarks for comparison. They're a bit mixed in some places, but it may just be because of difference in memory allocation when you don't reserve the number of columns in advance. It might be nice if we created an annotation for giving us the size of the table so that we can properly reserve the correct amount of memory, but I'm not sure how we would do that.

Either way, I think this is likely a net positive because it brings us closer to using exclusively arrow and it brings the total number of allocations down across the board so, pending CI and reapproval, I'll likely merge this anyway.

```
benchmark                                                          old ns/op     new ns/op     delta
BenchmarkFluxEndToEnd/set-12                                       17448414      17231816      -1.24%
BenchmarkFluxEndToEnd/check-12                                     304628258     303138138     -0.49%
BenchmarkFluxEndToEnd/deadman_add-12                               63257282      63272615      +0.02%
BenchmarkFluxEndToEnd/deadman_sub-12                               64493667      62676974      -2.82%
BenchmarkFluxEndToEnd/notify-12                                    129581834     127617050     -1.52%
BenchmarkFluxEndToEnd/state_changes-12                             155960229     155126207     -0.53%
BenchmarkFluxEndToEnd/secrets-12                                   14834427      14656088      -1.20%
BenchmarkFluxEndToEnd/measurement_tag_values-12                    40076767      39691869      -0.96%
BenchmarkFluxEndToEnd/show_measurements-12                         17954802      17893184      -0.34%
BenchmarkFluxEndToEnd/show_tag_keys-12                             17794335      17630692      -0.92%
BenchmarkFluxEndToEnd/show_tag_values-12                           18017733      17798653      -1.22%
BenchmarkFluxEndToEnd/replaceAll-12                                34460887      33979448      -1.40%
BenchmarkFluxEndToEnd/replace-12                                   35469533      34758541      -2.00%
BenchmarkFluxEndToEnd/subset-12                                    34841867      34157756      -1.96%
BenchmarkFluxEndToEnd/title-12                                     33017596      32410177      -1.84%
BenchmarkFluxEndToEnd/toLower-12                                   33098316      32735307      -1.10%
BenchmarkFluxEndToEnd/toUpper-12                                   33163131      32578308      -1.76%
BenchmarkFluxEndToEnd/aggregate_window_median-12                   189500174     186789360     -1.43%
BenchmarkFluxEndToEnd/buckets-12                                   13689918      13538727      -1.10%
BenchmarkFluxEndToEnd/combine_pivot-12                             16358460      16240644      -0.72%
BenchmarkFluxEndToEnd/delete_drop-12                               14323457      14493595      +1.19%
BenchmarkFluxEndToEnd/eval_map_with-12                             21804467      21404473      -1.83%
BenchmarkFluxEndToEnd/fill_default-12                              14224237      13984196      -1.69%
BenchmarkFluxEndToEnd/flatten_pivot-12                             16546163      16468322      -0.47%
BenchmarkFluxEndToEnd/noop_yield-12                                14670459      14230221      -3.00%
BenchmarkFluxEndToEnd/capitalize_strings_toUpper-12                34673548      33706246      -2.79%
BenchmarkFluxEndToEnd/cat_strings_joinStr-12                       35331065      34491788      -2.38%
BenchmarkFluxEndToEnd/center_strings_joinStr-12                    36807508      35834241      -2.64%
BenchmarkFluxEndToEnd/contains_filter_by_regex_match-12            28481495      26738185      -6.12%
BenchmarkFluxEndToEnd/count_string_countStr-12                     38843341      36278056      -6.60%
BenchmarkFluxEndToEnd/endswith_strings_regexp_hasSuffix-12         67181144      64610966      -3.83%
BenchmarkFluxEndToEnd/extract_regexp_findString-12                 37809539      36694969      -2.95%
BenchmarkFluxEndToEnd/get_strings_substring-12                     35070564      34152183      -2.62%
BenchmarkFluxEndToEnd/ljust_string_joinStr-12                      35556966      34519701      -2.92%
BenchmarkFluxEndToEnd/lower_strings_toLower-12                     33356914      32658933      -2.09%
BenchmarkFluxEndToEnd/lstrip_string_trimLeft-12                    34401132      33376604      -2.98%
BenchmarkFluxEndToEnd/repeat_strings_repeat-12                     34531807      33570717      -2.78%
BenchmarkFluxEndToEnd/replace_strings_replace-12                   35864833      35358332      -1.41%
BenchmarkFluxEndToEnd/rstrip_string_trimRight-12                   34457158      33500697      -2.78%
BenchmarkFluxEndToEnd/startswith_strings_regexp_hasPrefix-12       45564387      44253372      -2.88%
BenchmarkFluxEndToEnd/strip_strings_trimLeftRight-12               42127922      41331644      -1.89%
BenchmarkFluxEndToEnd/title_strings_title-12                       33563213      32759144      -2.40%
BenchmarkFluxEndToEnd/aggregate_empty_window-12                    15924134      15425141      -3.13%
BenchmarkFluxEndToEnd/aggregate_fill_window-12                     16972922      16487882      -2.86%
BenchmarkFluxEndToEnd/aggregate_window_max-12                      16980900      16836211      -0.85%
BenchmarkFluxEndToEnd/aggregate_window_median#01-12                19882644      19089355      -3.99%
BenchmarkFluxEndToEnd/aggregate_window-12                          17065078      16333560      -4.29%
BenchmarkFluxEndToEnd/cmo-12                                       15845529      15198996      -4.08%
BenchmarkFluxEndToEnd/columns-12                                   14050434      13473862      -4.10%
BenchmarkFluxEndToEnd/count-12                                     14219947      13635866      -4.11%
BenchmarkFluxEndToEnd/cov-12                                       42900231      41901352      -2.33%
BenchmarkFluxEndToEnd/covariance-12                                14646025      14104281      -3.70%
BenchmarkFluxEndToEnd/cumulative_sum_default-12                    14039781      13775436      -1.88%
BenchmarkFluxEndToEnd/cumulative_sum_noop-12                       14585816      13664230      -6.32%
BenchmarkFluxEndToEnd/cumulative_sum-12                            14909931      14379430      -3.56%
BenchmarkFluxEndToEnd/date_hour-12                                 29488547      27855100      -5.54%
BenchmarkFluxEndToEnd/date_microsecond-12                          29271665      27619153      -5.65%
BenchmarkFluxEndToEnd/date_millisecond-12                          29294790      28317572      -3.34%
BenchmarkFluxEndToEnd/date_minute-12                               29461232      27861189      -5.43%
BenchmarkFluxEndToEnd/date_month_day-12                            29561535      27832267      -5.85%
BenchmarkFluxEndToEnd/date_month-12                                29158515      27490061      -5.72%
BenchmarkFluxEndToEnd/date_nanosecond-12                           29129262      27576869      -5.33%
BenchmarkFluxEndToEnd/date_quarter-12                              29143946      27836048      -4.49%
BenchmarkFluxEndToEnd/date_second-12                               29342567      27748541      -5.43%
BenchmarkFluxEndToEnd/date_truncate-12                             31338385      29970409      -4.37%
BenchmarkFluxEndToEnd/date_week_day-12                             28939847      27652067      -4.45%
BenchmarkFluxEndToEnd/date_week-12                                 29488116      27623813      -6.32%
BenchmarkFluxEndToEnd/date_year_day-12                             29319334      27586417      -5.91%
BenchmarkFluxEndToEnd/date_year-12                                 29018115      27937144      -3.73%
BenchmarkFluxEndToEnd/derivative_nonnegative-12                    15000915      14315112      -4.57%
BenchmarkFluxEndToEnd/derivative-12                                14947857      13924142      -6.85%
BenchmarkFluxEndToEnd/difference_columns-12                        14918653      14327613      -3.96%
BenchmarkFluxEndToEnd/difference_keepfirst-12                      14897963      14054632      -5.66%
BenchmarkFluxEndToEnd/difference_nonnegative-12                    14859442      14060122      -5.38%
BenchmarkFluxEndToEnd/difference_one_value-12                      14591184      13927756      -4.55%
BenchmarkFluxEndToEnd/difference_panic-12                          21539971      20465741      -4.99%
BenchmarkFluxEndToEnd/difference-12                                14455441      13730522      -5.01%
BenchmarkFluxEndToEnd/distinct-12                                  15564587      15045750      -3.33%
BenchmarkFluxEndToEnd/double_exponential_moving_average-12         80509498      80765411      +0.32%
BenchmarkFluxEndToEnd/drop_after_rename-12                         15847270      15977322      +0.82%
BenchmarkFluxEndToEnd/drop_fn-12                                   21185273      21115170      -0.33%
BenchmarkFluxEndToEnd/drop_newname_after-12                        16448861      16711836      +1.60%
BenchmarkFluxEndToEnd/drop_newname_before-12                       16483822      16650728      +1.01%
BenchmarkFluxEndToEnd/drop_non_existent-12                         15806096      15918589      +0.71%
BenchmarkFluxEndToEnd/drop_unused-12                               22822928      22883947      +0.27%
BenchmarkFluxEndToEnd/duplicate_overwrite-12                       14983735      15364894      +2.54%
BenchmarkFluxEndToEnd/duplicate-12                                 15328307      15208687      -0.78%
BenchmarkFluxEndToEnd/dynamic_query-12                             36021959      37150090      +3.13%
BenchmarkFluxEndToEnd/elapsed_median-12                            34559715      35152376      +1.71%
BenchmarkFluxEndToEnd/elapsed-12                                   14459275      14938828      +3.32%
BenchmarkFluxEndToEnd/experimental_group-12                        28675272      28941570      +0.93%
BenchmarkFluxEndToEnd/exponential_moving_average-12                14526768      15044441      +3.56%
BenchmarkFluxEndToEnd/fill_bool-12                                 14847166      15190221      +2.31%
BenchmarkFluxEndToEnd/fill_float-12                                14921222      15141042      +1.47%
BenchmarkFluxEndToEnd/fill_int-12                                  15339602      15588054      +1.62%
BenchmarkFluxEndToEnd/fill_previous-12                             12272969      12327843      +0.45%
BenchmarkFluxEndToEnd/fill_string-12                               15370088      15107923      -1.71%
BenchmarkFluxEndToEnd/fill_time-12                                 14534722      14288534      -1.69%
BenchmarkFluxEndToEnd/fill_uint-12                                 15815203      15989068      +1.10%
BenchmarkFluxEndToEnd/filter_by_regex_compile-12                   25316208      25562594      +0.97%
BenchmarkFluxEndToEnd/filter_by_regex_function-12                  24717200      24908121      +0.77%
BenchmarkFluxEndToEnd/filter_by_regex-12                           23005786      23348311      +1.49%
BenchmarkFluxEndToEnd/filter_by_tags-12                            20941501      20919727      -0.10%
BenchmarkFluxEndToEnd/first-12                                     13094402      13340823      +1.88%
BenchmarkFluxEndToEnd/group_by_field-12                            15119689      15121331      +0.01%
BenchmarkFluxEndToEnd/group_by_irregular-12                        27078211      26869914      -0.77%
BenchmarkFluxEndToEnd/group_except-12                              16297923      16346070      +0.30%
BenchmarkFluxEndToEnd/group_nulls-12                               13910422      13981969      +0.51%
BenchmarkFluxEndToEnd/group-12                                     24366691      24048383      -1.31%
BenchmarkFluxEndToEnd/group_ungroup-12                             24556918      24373101      -0.75%
BenchmarkFluxEndToEnd/highestAverage-12                            19824372      19761717      -0.32%
BenchmarkFluxEndToEnd/highestCurrent-12                            21376936      21065896      -1.46%
BenchmarkFluxEndToEnd/highestMax-12                                19659742      19769459      +0.56%
BenchmarkFluxEndToEnd/histogram_normalize-12                       15076486      15193519      +0.78%
BenchmarkFluxEndToEnd/histogram_quantile_minvalue-12               14865096      15003397      +0.93%
BenchmarkFluxEndToEnd/histogram_quantile-12                        15731454      15511269      -1.40%
BenchmarkFluxEndToEnd/histogram-12                                 14263800      14282779      +0.13%
BenchmarkFluxEndToEnd/holt_winters_panic-12                        21141248      20846336      -1.39%
BenchmarkFluxEndToEnd/holt_winters-12                              84865198      83281070      -1.87%
BenchmarkFluxEndToEnd/hour_selection-12                            15100357      14898361      -1.34%
BenchmarkFluxEndToEnd/increase-12                                  16178776      16063606      -0.71%
BenchmarkFluxEndToEnd/influxFieldsAsCols-12                        21249976      22109798      +4.05%
BenchmarkFluxEndToEnd/influxRawQuery-12                            53326420      53553364      +0.43%
BenchmarkFluxEndToEnd/integral-12                                  13915965      13717054      -1.43%
BenchmarkFluxEndToEnd/join_across_measurements-12                  56021890      55993853      -0.05%
BenchmarkFluxEndToEnd/join_agg-12                                  49645696      48901915      -1.50%
BenchmarkFluxEndToEnd/join-12                                      43461511      42136953      -3.05%
BenchmarkFluxEndToEnd/kama-12                                      16197577      16316488      +0.73%
BenchmarkFluxEndToEnd/kama_v2-12                                   50382182      50468865      +0.17%
BenchmarkFluxEndToEnd/keep_fn-12                                   29494293      29071612      -1.43%
BenchmarkFluxEndToEnd/keep_non_existent_only-12                    15001816      15079966      +0.52%
BenchmarkFluxEndToEnd/keep_non_existent-12                         16074295      16039263      -0.22%
BenchmarkFluxEndToEnd/keep-12                                      15512361      15513253      +0.01%
BenchmarkFluxEndToEnd/ker-12                                       83958914      83049176      -1.08%
BenchmarkFluxEndToEnd/key_values_host_name-12                      13570594      13547814      -0.17%
BenchmarkFluxEndToEnd/key_values-12                                14077633      13680668      -2.82%
BenchmarkFluxEndToEnd/keys-12                                      14626597      14990448      +2.49%
BenchmarkFluxEndToEnd/last-12                                      13747635      13485667      -1.91%
BenchmarkFluxEndToEnd/limit_offset-12                              15516172      15597735      +0.53%
BenchmarkFluxEndToEnd/limit-12                                     15216565      15458037      +1.59%
BenchmarkFluxEndToEnd/lowestAverage-12                             21482086      21682600      +0.93%
BenchmarkFluxEndToEnd/lowestCurrent-12                             21814773      21506518      -1.41%
BenchmarkFluxEndToEnd/lowestMin-12                                 21646418      21581377      -0.30%
BenchmarkFluxEndToEnd/map_extension_with-12                        23834524      23485332      -1.47%
BenchmarkFluxEndToEnd/map_extern_dynamic_var-12                    31004147      30901585      -0.33%
BenchmarkFluxEndToEnd/map_extern_var-12                            30614091      30082093      -1.74%
BenchmarkFluxEndToEnd/map_local_var-12                             29433278      30465613      +3.51%
BenchmarkFluxEndToEnd/map_polymorphism-12                          23761036      24529304      +3.23%
BenchmarkFluxEndToEnd/map_shadow_var-12                            31818292      31604687      -0.67%
BenchmarkFluxEndToEnd/map-12                                       20346586      21009864      +3.26%
BenchmarkFluxEndToEnd/map_with_obj-12                              51574974      52016418      +0.86%
BenchmarkFluxEndToEnd/math_fn_constant-12                          63583737      64054532      +0.74%
BenchmarkFluxEndToEnd/math_m_max-12                                70556358      69166179      -1.97%
BenchmarkFluxEndToEnd/max-12                                       13782521      13725466      -0.41%
BenchmarkFluxEndToEnd/max_time-12                                  15324001      15440242      +0.76%
BenchmarkFluxEndToEnd/mean-12                                      14427127      14775204      +2.41%
BenchmarkFluxEndToEnd/median_column-12                             15407039      15146445      -1.69%
BenchmarkFluxEndToEnd/median_compression-12                        14856694      15017219      +1.08%
BenchmarkFluxEndToEnd/median_method-12                             15349604      15055702      -1.91%
BenchmarkFluxEndToEnd/median-12                                    15444863      15671880      +1.47%
BenchmarkFluxEndToEnd/meta_query_keys-12                           22506646      22659380      +0.68%
BenchmarkFluxEndToEnd/min-12                                       13531141      13547540      +0.12%
BenchmarkFluxEndToEnd/mode_string-12                               14294256      14499975      +1.44%
BenchmarkFluxEndToEnd/mode-12                                      14592697      14345617      -1.69%
BenchmarkFluxEndToEnd/moving_average-12                            14912005      14913159      +0.01%
BenchmarkFluxEndToEnd/multiple_range-12                            15213624      15326262      +0.74%
BenchmarkFluxEndToEnd/parse_regex-12                               22984743      22622965      -1.57%
BenchmarkFluxEndToEnd/pivot_col_order-12                           17231130      16910703      -1.86%
BenchmarkFluxEndToEnd/pivot_fields-12                              17152187      17042401      -0.64%
BenchmarkFluxEndToEnd/pivot_mean-12                                19611507      19498783      -0.57%
BenchmarkFluxEndToEnd/pivot_task_test-12                           25181641      24572947      -2.42%
BenchmarkFluxEndToEnd/pivot-12                                     17341122      17116456      -1.30%
BenchmarkFluxEndToEnd/promql_changes-12                            194690875     191467694     -1.66%
BenchmarkFluxEndToEnd/promql_dayOfMonth-12                         372127197     375017344     +0.78%
BenchmarkFluxEndToEnd/promql_dayOfWeek-12                          369611343     381001936     +3.08%
BenchmarkFluxEndToEnd/promql_daysInMonth-12                        376565974     378872386     +0.61%
BenchmarkFluxEndToEnd/promql_emptyTable-12                         153041974     151613228     -0.93%
BenchmarkFluxEndToEnd/promql_extrapolatedRate_counter_rate-12      210248947     200397421     -4.69%
BenchmarkFluxEndToEnd/promql_extrapolatedRate_nocounter-12         208508983     196133918     -5.94%
BenchmarkFluxEndToEnd/promql_extrapolatedRate_norate-12            202528019     202031461     -0.25%
BenchmarkFluxEndToEnd/promql_histogramQuantile-12                  196926519     198139264     +0.62%
BenchmarkFluxEndToEnd/promql_holtWinters-12                        215465794     204607767     -5.04%
BenchmarkFluxEndToEnd/promql_hour-12                               423088916     377443798     -10.79%
BenchmarkFluxEndToEnd/promql_instantRate-12                        200815488     201947779     +0.56%
BenchmarkFluxEndToEnd/promql_labelReplace_empty_dst-12             196409950     195584696     -0.42%
BenchmarkFluxEndToEnd/promql_labelReplace_full_string_match-12     194936485     194375217     -0.29%
BenchmarkFluxEndToEnd/promql_labelReplace_multiple_groups-12       192397755     193336304     +0.49%
…